+++ app/showcase-qa/page.tsx
@@ -1,0 +1,5 @@
+import { ShowcaseQaDemoSmoke } from "@/components/ShowcaseQaDemoSmoke";
+
+export default function Page() {
+  return <ShowcaseQaDemoSmoke />;
+}